# backend/api_docs.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Breathometer 4.0 Backend",
    "description": "Real-time backend API for Breathometer 4.0 health assessment platform.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/environment": {
      "post": {
        "tags": [
          "environment"
        ],
        "summary": "Store Environment Data",
        "operationId": "store_environment_data_environment_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnvironmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/environment/aqi": {
      "get": {
        "tags": [
          "environment"
        ],
        "summary": "Get Aqi",
        "description": "Fetch real AQI for a lat/lon coordinate via AQICN.",
        "operationId": "get_aqi_environment_aqi_get",
        "parameters": [
          {
            "name": "lat",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "default": 0.0,
              "title": "Lat"
            }
          },
          {
            "name": "lon",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "default": 0.0,
              "title": "Lon"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/environment/aqi-by-city": {
      "get": {
        "tags": [
          "environment"
        ],
        "summary": "Get Aqi By City",
        "description": "Fetch real AQI for a city name via AQICN.",
        "operationId": "get_aqi_by_city_environment_aqi_by_city_get",
        "parameters": [
          {
            "name": "city",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "City"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/environment/weather": {
      "get": {
        "tags": [
          "environment"
        ],
        "summary": "Get Weather",
        "description": "Fetch real weather data for a lat/lon coordinate via OpenWeatherMap.",
        "operationId": "get_weather_environment_weather_get",
        "parameters": [
          {
            "name": "lat",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "default": 0.0,
              "title": "Lat"
            }
          },
          {
            "name": "lon",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "default": 0.0,
              "title": "Lon"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/prediction/store": {
      "post": {
        "tags": [
          "prediction"
        ],
        "summary": "Store Prediction",
        "operationId": "store_prediction_prediction_store_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/app__routes__prediction__PredictionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/prediction/{user_id_param}": {
      "get": {
        "tags": [
          "prediction"
        ],
        "summary": "Get Prediction History",
        "operationId": "get_prediction_history_prediction__user_id_param__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id_param",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id Param"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EnvironmentRequest": {
        "properties": {
          "pm25": {
            "type": "number",
            "title": "Pm25"
          },
          "pm10": {
            "type": "number",
            "title": "Pm10"
          },
          "aqi": {
            "type": "number",
            "title": "Aqi"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          }
        },
        "type": "object",
        "required": [
          "pm25",
          "pm10",
          "aqi"
        ],
        "title": "EnvironmentRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "app__routes__prediction__PredictionRequest": {
        "properties": {
          "final_risk_score": {
            "type": "number",
            "title": "Final Risk Score",
            "default": 0.0
          },
          "risk_category": {
            "type": "string",
            "title": "Risk Category"
          },
          "ai_explanation": {
            "type": "string",
            "title": "Ai Explanation"
          },
          "top_risk_factors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Top Risk Factors"
          }
        },
        "type": "object",
        "required": [
          "risk_category",
          "ai_explanation",
          "top_risk_factors"
        ],
        "title": "PredictionRequest"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}